Почистила код
перенесла addToFavorite()
функцію роботи чекбоксів в окремий файл для зручності читання коду
